New versatile Menu
The navigation menu can now resize depending on screen width. It will
modify itself when the screen changes to certain sizes so that it is
still visible and more easily accessible. One example is if the browser
is the size of a smart phone it will have a drop down menu.

I also made the menu highlight the page you are on instead of always
highlighting the Home..